Add files via upload
This code uses the Deribit API to retrieve historical price data for ETH and retrieves the current price of ETH in USD. It then uses the historical data to calculate the moving average of the last 50 data points, and compares it with the current price to determine whether it is an appropriate time to buy. If the current price is below the moving average and the budget allows, the code places a buy order using the place_order function. If the current price is above the moving average or the budget is not enough, it will print a message.